pyproject.toml
[tool.poetry]
name = "bnex-test"
version = "0.1.0"
description = ""
authors = ["João Pedro Prado <joaopedro0128@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django = "^4.2.7"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
ruff = "^0.1.3"
black = "^23.10.1"
pytest = "^7.4.3"
pytest-django = "^4.6.0"
faker = "^19.12.1"
pre-commit = "^3.5.0"
model-bakery = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=88
target-version=['py310']
extend-exclude='''
(
  ^/manage.py
  | ^/data/
  | ^/bnex/
  | migrations
)
'''

[tool.isort]
profile="black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="bnex.settings"
python_files = ["test_*.py"]
addopts = "--ignore=data"